Separate db and db provider
